requirements.txt
streamlit
PyMuPDF
google-genai